requirements.txt
django==4.0.6
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-localflavor ==3.1
Pillow==9.2.0